Correct projected and monthly savings from float to int

+++ python_introduction/finance_calculator.py
@@ -22,10 +22,10 @@ monthly_income = float(input("Enter your monthly income: "))
 total_monthly_expenses = float(input("Enter your total monthly expenses: "))
 
 # Calculate monthly savings
-monthly_savings = monthly_income - total_monthly_expenses
+monthly_savings = int(monthly_income - total_monthly_expenses)
 
 # Project Annual Savings:
-projected_savings = (monthly_savings * 12) + (monthly_savings * 12 * 0.05)
+projected_savings = int((monthly_savings * 12) + (monthly_savings * 12 * 0.05))
 
 # Display the user’s monthly savings.
 # Display the projected annual savings after including interest.
